+++ scripts/render.js
@@ -5,17 +5,14 @@ ui,
 renderTemplate,
 AmbientSoundConfig,
 AmbientLightConfig,
-TokenConfig
+TokenConfig,
+DefaultTokenConfig
 */
 
 "use strict";
 
 import { log } from "./module.js";
 import { KEYS, MODULE_ID, TEMPLATES, HTML_INJECTION } from "./const.js";
-import {
-  lightMaskUpdateCustomEdgeCache,
-  lightMaskShiftCustomEdgeCache } from "./preUpdate.js";
-
 
 /**
  * Inject new template information into the configuration render
@@ -86,79 +83,6 @@ async function injectConfiguration(app, html, data, type) {
 }
 
 /**
- * Wrap activateListeners to catch when user clicks the button to add custom wall ids.
- */
-export function lightMaskActivateListeners(wrapped, html) {
-  log(`lightMaskActivateListeners html[0] is length ${html[0].length}`, html, this);
-
-  html.on("click", ".saveWallsButton", onAddWallIDs.bind(this));
-  html.on("click", ".lightmaskRelativeCheckbox", onCheckRelative.bind(this));
-
-  return wrapped(html);
-}
-
-/**
- * Add a method to the AmbientLightConfiguration to handle when user
- * clicks the button to add custom wall ids.
- * @param {PointerEvent} event    The originating click event
- */
-function onAddWallIDs(event) {
-  log("lightMaskOnAddWallIDs", event, this);
-
-  const ids_to_add = controlledWallIDs();
-  if (!ids_to_add) return;
-  log(`Ids to add: ${ids_to_add}`);
-
-  // Change the data and refresh...
-  let edges_cache = this.object.getFlag(MODULE_ID, KEYS.CUSTOM_WALLS.EDGES) || [];
-  edges_cache = lightMaskUpdateCustomEdgeCache(edges_cache, ids_to_add);
-
-  const newData = {
-    [`flags.${MODULE_ID}.${KEYS.CUSTOM_WALLS.IDS}`]: ids_to_add,
-    [`flags.${MODULE_ID}.${KEYS.CUSTOM_WALLS.EDGES}`]: edges_cache
-  };
-
-  const previewData = this._getSubmitData(newData);
-  foundry.utils.mergeObject(this.object.data, previewData, {inplace: true});
-
-  this.render();
-}
-
-/**
- * Listener to handle when a user check/unchecks the "Relative" checkbox.
- * If "Relative" is checked, the edges cache must be updated by a directional vector
- * based on the shift in origin.
- * @param {PointerEvent} event    The originating click event
- */
-function onCheckRelative(event) {
-  log("lightMaskOnCheckRelative", event, this);
-
-  const current_origin = { x: this.object.data.x,
-                           y: this.object.data.y }; // eslint-disable-line indent
-  const newData = {};
-  if (event.target.checked) {
-    // Update with the new origin
-    newData[`flags.${MODULE_ID}.${KEYS.ORIGIN}`] = current_origin;
-
-  } else {
-    // Set the wall locations based on the last origin because when the user unchecks
-    // relative, we want the walls to stay at the last relative position (not their
-    // original position)
-    let edges_cache = this.object.getFlag(MODULE_ID, KEYS.CUSTOM_WALLS.EDGES) || [];
-    const stored_origin = this.object.getFlag(MODULE_ID, KEYS.ORIGIN) || current_origin;
-    const delta = { dx: current_origin.x - stored_origin.x,
-                    dy: current_origin.y - stored_origin.y }; // eslint-disable-line indent
-
-    edges_cache = lightMaskShiftCustomEdgeCache(edges_cache, delta);
-    newData[`flags.${MODULE_ID}.${KEYS.CUSTOM_WALLS.EDGES}`] = edges_cache;
-  }
-
-  const previewData = this._getSubmitData(newData);
-  foundry.utils.mergeObject(this.object.data, previewData, {inplace: true});
-  this.render();
-}
-
-/**
  * If the shape rotation has changed, update flags so the UI can be updated accordingly.
  * Only relevant for AmbientSoundConfig and TokenConfig. AmbientLightConfig already
  * changes rotation.
@@ -166,15 +90,15 @@ function onCheckRelative(event) {
 export async function updateRotation(event) {
   log("updateRotation", event, this);
 
-  let docData = this.document?.data;
-  if ( this instanceof TokenConfig ) docData = this.token.data;
+  let doc = this.document;
+  if ( this instanceof TokenConfig ) doc = this.token;
 
   const rotation = parseInt(event.target.value);
   const newData = {};
   newData[`flags.${MODULE_ID}.${KEYS.ROTATION}`] = rotation;
 
   const previewData = this._getSubmitData(newData);
-  foundry.utils.mergeObject(docData, previewData, {inplace: true});
+  foundry.utils.mergeObject(doc, previewData, {inplace: true});
 }
 
 /**
@@ -187,17 +111,17 @@ export async function updateShapeIndicator(event) {
   log("updateShapeIndicator", event, this);
 
   let doc = this.document;
-  let docData = this.document?.data;
 
   if ( this instanceof DefaultTokenConfig ) {
     log("Default token data update");
     doc = this.token;
-    docData = this.data;
+//     docData = this.data;
 
   } else if ( this instanceof TokenConfig ) {
     log("Token data update");
-    doc = this.token;
-    docData = this.isPrototype ? this.actor.data.token : this.token.data;
+    doc = this.isPrototype ? this.actor.token : this.token;
+//     doc = this.token;
+//     docData = this.isPrototype ? this.actor.data.token : this.token.data;
   }
 
   const shape = event.target.value;
@@ -214,11 +138,11 @@ export async function updateShapeIndicator(event) {
 
     let major = 0;
     if ( this instanceof AmbientSoundConfig ) {
-      major = docData.radius;
+      major = doc.radius;
     } else if ( this instanceof AmbientLightConfig ) {
-      major = Math.max(docData.config.dim, docData.config.bright);
+      major = Math.max(doc.config.dim, doc.config.bright);
     } else if ( this instanceof TokenConfig ) {
-      major = Math.max(docData.light.dim, docData.light.bright);
+      major = Math.max(doc.light.dim, doc.light.bright);
     } else {
       console.warn("updateShapeIndicator|Config object not recognized.", this);
     }
@@ -229,21 +153,5 @@ export async function updateShapeIndicator(event) {
   }
 
   const previewData = this._getSubmitData(newData);
-  foundry.utils.mergeObject(docData, previewData, {inplace: true});
-}
-
-/**
- * Retrieve a comma-separated list of wall ids currently controlled on the canvas.
- * @return {string}
- */
-export function controlledWallIDs() {
-  const walls = canvas.walls.controlled;
-  if (walls.length === 0) {
-    console.warn("Please select one or more walls on the canvas.");
-    ui.notifications.warn("Please select one or more walls on the canvas.");
-    return;
-  }
-
-  const id = walls.map(w => w.id);
-  return id.join(",");
+  foundry.utils.mergeObject(doc, previewData, {inplace: true});
 }